Major refactor on Domain object and DTO and processing

+++ myfinAPI/Model/DTO/AssetClass.cs
@@ -8,9 +8,9 @@ namespace myfinAPI.Model.DTO
 	public class AssetClass
 	{
 		public string AssetClassName { get; set; }
-		public double Investment { get; set; }
+		public Decimal Investment { get; set; }
 		public double Profit { get; set; }
-		public double percent{ get; set; }
+		public decimal percent{ get; set; }
 		public CompanySize cmpSize { get; set; }
 	}
 	public enum CompanySize
