+++ myfinAPI/Model/DTO/AssetReturn.cs
@@ -8,9 +8,9 @@ namespace myfinAPI.Model.DTO
 	public class AssetReturn
 	{
 		public int PortfolioId { get; set; }
-		public double Return { get; set; }
-		public double xirr { get; set; }
+		public decimal Return { get; set; }
+		public decimal xirr { get; set; }
 		public int year { get; set; }
-		public double dividend { get; set; }
+		public decimal dividend { get; set; }
 	}
 }
