+++ myfinAPI/Model/Domain/Cashflow.cs
@@ -9,8 +9,8 @@ namespace myfinAPI.Model.DTO
 	public class CashFlow
 	{
 		public int portfolioId { get; set; }
-		public double Cashflow { get; set; }
-		public double Dividend { get;set; }		
+		public decimal Cashflow { get; set; }
+		public decimal Dividend { get;set; }		
 		public int month { get; set; }
 		public int year { get; set; }		
 		public AssetType Assettype { get; set; }
